Do not execute vnet-admin-setup if tsh was built with daemon support
