+++ tool/tsh/common/vnet_darwin.go
@@ -21,17 +21,12 @@ package common
 
 import (
 	"fmt"
-	"log/slog"
-	"os"
 
 	"github.com/alecthomas/kingpin/v2"
 	"github.com/gravitational/trace"
 
 	"github.com/gravitational/teleport"
-	"github.com/gravitational/teleport/api/types"
-	"github.com/gravitational/teleport/lib/utils"
 	"github.com/gravitational/teleport/lib/vnet"
-	"github.com/gravitational/teleport/lib/vnet/daemon"
 )
 
 type vnetCommand struct {
@@ -100,28 +95,6 @@ func newVnetAdminSetupCommand(app *kingpin.Application) *vnetAdminSetupCommand {
 	return cmd
 }
 
-func (c *vnetAdminSetupCommand) run(cf *CLIConf) error {
-	homePath := os.Getenv(types.HomeEnvVar)
-	if homePath == "" {
-		// This runs as root so we need to be configured with the user's home path.
-		return trace.BadParameter("%s must be set", types.HomeEnvVar)
-	}
-
-	config := daemon.Config{
-		SocketPath: c.socketPath,
-		IPv6Prefix: c.ipv6Prefix,
-		DNSAddr:    c.dnsAddr,
-		HomePath:   homePath,
-		ClientCred: daemon.ClientCred{
-			Valid: true,
-			Egid:  c.egid,
-			Euid:  c.euid,
-		},
-	}
-
-	return trace.Wrap(vnet.AdminSetup(cf.Context, config))
-}
-
 type vnetDaemonCommand struct {
 	*kingpin.CmdClause
 	// Launch daemons added through SMAppService are launched from a static .plist file, hence
@@ -131,17 +104,9 @@ type vnetDaemonCommand struct {
 
 func newVnetDaemonCommand(app *kingpin.Application) *vnetDaemonCommand {
 	return &vnetDaemonCommand{
-		// The command must match the command provided in the .plist file.
-		CmdClause: app.Command("vnet-daemon", "Start the VNet daemon").Hidden(),
+		CmdClause: app.Command(vnetDaemonSubCommand, "Start the VNet daemon").Hidden(),
 	}
 }
 
-func (c *vnetDaemonCommand) run(cf *CLIConf) error {
-	if cf.Debug {
-		utils.InitLogger(utils.LoggingForDaemon, slog.LevelDebug)
-	} else {
-		utils.InitLogger(utils.LoggingForDaemon, slog.LevelInfo)
-	}
-
-	return trace.Wrap(vnet.DaemonSubcommand(cf.Context))
-}
+// The command must match the command provided in the .plist file.
+const vnetDaemonSubCommand = "vnet-daemon"
